move the interop tests to a subdirectory
