+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "integrationtests/qifs"]
-	path = integrationtests/qifs
+[submodule "integrationtests/interop/qifs"]
+	path = integrationtests/interop/qifs
 	url = git@github.com:qpackers/qifs.git
