+++ integrationtests/interop/interop_test.go
@@ -1,4 +1,4 @@
-package integrationtests
+package interop
 
 import (
 	"encoding/binary"
